src/main/java/com/example/proceedingsmanagement/business/rabbit/MessageSender.java
package com.example.proceedingsmanagement.business.rabbit;

import org.springframework.amqp.core.MessageDeliveryMode;
import org.springframework.amqp.rabbit.core.RabbitTemplate;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Service;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

@Service
public class MessageSender {

    private static final Logger logger = LoggerFactory.getLogger(MessageSender.class);

    private final RabbitTemplate rabbitTemplate;

    @Value("${rabbitmq.exchange}")
    private String exchange;

    @Value("${rabbitmq.routingkey}")
    private String routingKey;

    public MessageSender(RabbitTemplate rabbitTemplate) {
        this.rabbitTemplate = rabbitTemplate;
    }

    public boolean sendMessage(String messageData) {
        try {
            rabbitTemplate.convertAndSend(exchange, routingKey, messageData, message -> {
                message.getMessageProperties().setDeliveryMode(MessageDeliveryMode.PERSISTENT);
                return message;
            });
            logger.info("Message Sent: {}", messageData);
            return true; 
        } catch (Exception e) {
            logger.error("Failed to send message: {}", messageData, e);
            return false;
        }
    }
}
